Use unidecode to get the first letter for the index
